requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
python-decouple>=3.8
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
openai>=1.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
django-cors-headers>=4.3.0

